네이버 간편 로그인 기능 추가 및 시간 치환 서비스 생성 (#18)
* Feat: 레시피 추천 버그 보완

1. 사용자가 등록한 식재료가 없을 때 예외 추가

* Feat: 레시피 검색 버그 수정 & 레시피 검색 조건 조회 기능 추가

1. 레시피 검색 버그 수정
-> 별점이 0.0 미만 5.0 초과일 경우, 에러 발생하도록 코드 수정
-> 조건이 null로 들어왔을 때, 동적쿼리가 제대로 작성되도록 코드 수정

2. 레시피 조건 조회 기능 추가
-> 레시피 유형 조회 기능 추가
-> 레시피 난이도 조회 기능 추가

* Feat: 구글 간편 로그인 기능 구현

* Feat: 최근 검색어 기능 추가

* Feat: 추천 검색어 기능 추가

* Feat: 식재료 차감 기능 추가

* Feat: 네이버 간편 로그인 기능 추가 및 시간 치환 서비스 생성

+++ back/src/main/java/refrigerator/back/ingredient/application/port/in/ModifyIngredientUseCase.java
@@ -6,6 +6,6 @@ import java.time.LocalDateTime;
 public interface ModifyIngredientUseCase {
 
     void modifyIngredient(Long id, LocalDate expirationDate,
-                          Integer capacity, String storageMethod);
+                          Double capacity, String storageMethod);
 
 }
